Update README.md
This code is implements Iterative Deepening Depth-First Search (IDDFS) to find a path in a 2D grid maze. The maze consists of empty spaces (0) and walls (1) and the task is to find if a valid path exists from a start point to a target point. Movement is allowed in four directions: up, down, left and right.